+++ examples/with-parcel/index.js
@@ -2,15 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Button } from 'rsuite';
 import './styles/index.less';
-class HelloMessage extends React.Component {
-  render() {
-    return (
-      <div>
-        <p>Hello {this.props.name}</p>
-        <Button appearance="primary">test</Button>
-      </div>
-    );
-  }
+
+function HelloMessage({ name }) {
+  return (
+    <div>
+      <p>Hello {name}</p>
+      <Button appearance="primary">test</Button>
+    </div>
+  );
 }
 
 var mountNode = document.getElementById('app');