+++ src/Dropdown/test/DropdownSpec.js
@@ -281,14 +281,44 @@ describe('<Dropdown>', () => {
 
   it('Should not call onToggle callback when set disabled', () => {
     const onToggleSpy = sinon.spy();
-    const instance = getDOMNode(
+    const { getByRole } = render(
       <Dropdown onToggle={onToggleSpy} disabled>
         <Dropdown.Item eventKey={1}>1</Dropdown.Item>
         <Dropdown.Item eventKey={2}>2</Dropdown.Item>
       </Dropdown>
     );
-    fireEvent.click(instance.querySelector('.rs-dropdown-toggle'));
-    assert.ok(!onToggleSpy.calledOnce);
+
+    fireEvent.click(getByRole('button'));
+
+    expect(onToggleSpy).to.have.not.been.called;
+  });
+
+  it('Should not call onToggle callback when set disabled and hover', () => {
+    const onToggleSpy = sinon.spy();
+    const { getByRole } = render(
+      <Dropdown onToggle={onToggleSpy} disabled trigger="hover">
+        <Dropdown.Item eventKey={1}>1</Dropdown.Item>
+        <Dropdown.Item eventKey={2}>2</Dropdown.Item>
+      </Dropdown>
+    );
+
+    fireEvent.mouseEnter(getByRole('button'));
+
+    expect(onToggleSpy).to.have.not.been.called;
+  });
+
+  it('Should not call onToggle callback when set disabled and contextMenu', () => {
+    const onToggleSpy = sinon.spy();
+    const { getByRole } = render(
+      <Dropdown onToggle={onToggleSpy} disabled trigger="contextMenu">
+        <Dropdown.Item eventKey={1}>1</Dropdown.Item>
+        <Dropdown.Item eventKey={2}>2</Dropdown.Item>
+      </Dropdown>
+    );
+
+    fireEvent.contextMenu(getByRole('button'));
+
+    expect(onToggleSpy).to.have.not.been.called;
   });
 
   it('Should have a custom style in Menu', () => {